Clone repo at startup if needed.
This is so that the whole plexconnect directory can be in a volume.
It's unfortunate, but since the important ATVSetting.cfg file is in
the root, that's the best option. It would be tiresome to loose Plex
login on every restart, which would happen if merely the certificates
were stored.